migrating the multiple file frame into a separate class in it's own file.
